api/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml>=5.3.0
googlemaps==4.10.0
phonenumbers==8.13.19
regex==2023.10.3
openai>=1.3.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
